Current version is 3.1.0 with 4.x in pre-release
The 4.x release has been moved to pre-release status as additional default pin changes are planned to resolve long-standing library conflicts with the servo library. A new branch for the 3.x version with the original default pins was created for the 3.1.0 release, and 3.x maintenance until the 4.0.x release is made.